stricter checking of types
This patchset addresses #39.

we've now added the entire (default) config of flake8-mypy.
The following options deviate from their default (which gets us a step
closer to mypy --strct)

- ignore_missing_imports=False
- follow_imports=True
- cache_dir=.mypy_cache
- disallow_untyped_calls=True
- warn_return_any=True

We've also toggled these two:

- warn_redundant_casts=True
- warn_unused_ignores=True

Finally, we check for

- check_untyped_defs=True

For the full --strict set, we'd also need:

- disallow_untyped_defs=True

But that, we'll need better typedefs.