Resource is an abstract class, be clear about it
the consequence of this is that we need to declare the non-existent
function destroy in `Resource`.
As a consequence, we have to explicitly declare it in the slightly
specialized classes that directly inherit from it, `DefaultResource`
and `ListableResource`. We "implement" it by raising an error.

This patch fixes #156, and has been once again discovered by `mypy
--strict` (#131).