make ioc start --rc idempotent
only filter the jails that aren't running already,
and don't fail when there aren't any.

The latter is done by making start_jails(): return success, instead of
exiting itself. this way we can loop the return value back to the caller,
and exit in one place, rather than doing it in a deep function call
that's shared among other methods of starting.

This addresses @gronke's feedack & will be applied to stop.py as well!