multiple IocageEvent history scopes for parallelization
Before this changes all IocageEvent instances shared a (singleton) data structure to keep track of the event history. With this change all public methods that yield events will either create a new scope or can be called with a reference to an existing scope. This allows us to run events of the same kind in parallel.