attempt to claim /iocage mountpoint on root dataset creation
When the zfs pool has no mountpoint set, iocage cannot determine the
the correct mountpoint. In tradition of iocage_legacy it is attempted
to claim /iocage if that location did not exist yet. In case it does
a ZFSSourceCreation error is raised reminding the user to manually
create the dataset and giving it a mountpoint.