Fix pkg install and freebsd update
during the review of #602, we decided to rename the Jail class variable
`ignore_user_defaults` to `ignore_source_config`, so as to not be confused
with `user.*` properties.

This was, however, not done consequentially throughout the PR, so that
`ioc pkg` and `ioc update` were still broken.

This PR fixes that.